refactor(#70): 만료 알림 코드 api 서버로 코드 분리
