+++ src/main/java/com/couponmoa/backend/domain/notification/enums/NotificationType.java
@@ -1,4 +1,4 @@
-package com.couponmoa.backend.notification.enums;
+package com.couponmoa.backend.domain.notification.enums;
 
 public enum NotificationType {
     EXPIRE_SOON, COUPON_SUBSCRIBE, STORE_SUBSCRIBE
