+++ src/main/java/com/couponmoa/backend/scheduler/notification/NotificationScheduler.java
@@ -1,6 +1,6 @@
 package com.couponmoa.backend.scheduler.notification;
 
-import com.couponmoa.backend.notification.service.NotificationService;
+import com.couponmoa.backend.domain.notification.service.NotificationService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
